+++ demo/log_to_file/endpoints/rpg/story/add.py
@@ -1,0 +1,14 @@
+"""
+Append a story to our rpg
+"""
+
+def post(req, api):
+    """
+    Input:
+        content: string
+        
+    Output:
+        string
+    """
+    api.debug(req.params['content'])
+    return 'Success'