+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# v0.3.0
+Added `--config` to command line to select config
+
+Added `log` options to config, now it's possible to log to a file
+
 ## v0.2.3
 Changed `req.param` to `req.params`